requirements.txt
# 🏷️ PriceChecker - Dépendances vérifiées et activement utilisées

# === FRAMEWORK WEB (OBLIGATOIRE) ===
Flask==3.1.1
Flask_moment==1.0.6

# === WEB SCRAPING (CORE FEATURES) ===
requests==2.32.4
beautifulsoup4==4.13.4
selenium==4.33.0
fake-useragent==2.2.0

# === ENVIRONNEMENT VIRTUEL (OBLIGATOIRE) ===
python-dotenv==1.1.0


# === OPTIONNEL ===
# === AMÉLIORATION DU PARSING HTML ===
# html5lib==1.1 # Changer le parser par défaut de BeautifulSoup dans price_scraper.py de 'html.parser' vers 'html5lib' (recommandé pour une meilleure compatibilité avec HTML5)

# === TESTS ===
# pytest==8.4.0
# pytest-cov==6.2.0

# === MODULES NATIFS PYTHON (PAS BESOIN D'INSTALLATION) ===
# sqlite3 - Base de données (inclus avec Python)
# csv - Export des données (inclus avec Python)  
# datetime - Gestion des dates (inclus avec Python)
# logging - Logs de l'application (inclus avec Python)
# re - Expressions régulières (inclus avec Python)
# hashlib - Hachage (inclus avec Python)
# json - Manipulation JSON (inclus avec Python)
# io.StringIO - Manipulation de chaînes (inclus avec Python)
# os - Système d'exploitation (inclus avec Python)

# === SÉCURITÉ (recommandé pour production) ===
# flask-wtf==1.2.2          # Si vous ajoutez protection CSRF
# flask-talisman==1.1.0      # Si vous ajoutez headers sécurité

# === PERFORMANCE (recommandé pour production) ===
# flask-compress==1.17       # Si vous ajoutez compression gzip
# gunicorn==23.0.0   # Pour déploiement production

# === DÉVELOPPEMENT ===
# vulture==2.14            # Pour détecter le code mort
# black==25.1.0            # Pour formatter le code
# flake8==7.2.0            # Pour linter le code